initial commit
This commit adds the erlang sourcecode of a very basic eIM
implementation. The eIM currently uses a lot of hardcoded parameters and
only works in a very specific test setup that conists of onomondo-ipad
and osmo-smdpp.

The eIM also comes with the implementation of a REST API, which is
connected to an mnesia database. This REST API is not yet connected to
the eIM code. The REST API can be tried out using tryme.py.