esipa_asn1_handler: use transactionId as second work identifier
At the moment we only use the PID of the HTTP connection to identify the
work item where we store the procedure state (e.g. the URL of the SMDP+).
This may not be enough. In case the HTTP connection breaks down it may
be re-estabished under a new PID. The next request may then come in
under a different PID and we won't be able to identify the related work
item. However, in case the requests contains a transactionId, we can use
that as an alternative identifier.