rest_handler: add facility name as a prefix to the order field
The returned JSON string contains a field named "order". The contents
of this field depend on the facility. This makes it hard to describe
in an abstract notation. Let's add the facility name as prefix.
("order" -> "download_order"). Then it is always clear for the receiver
what field contents to expect and the description in an abstract syntax
beomes much easier.

Related: SYS#6720