rest_handler: convey useful outcome back to the REST API user
At the moment the REST API user has to refer to the "outcome" field
in the returned JSON. Unfortunately this field is encoded in erlang's
ETF format, which requires a parser on the other end. This might work
for debugging but we cannot expect a REST API user to do complicated
parsing.

To fix this we should put useful JSON content into the outcome field
and add another field called "debuginfo" where we put the ETF
encoded information.

With this patche we already get useful information on download results
and return codes for most of the PSMO/ECOs. However, there are still
some TODOs here and there, which will be adressed in follow-up patches.

Related: SYS#6720