mnesia_db: store eUICC master data and handle counterValue
We currently have no persistent storage to store eUICC master data,
such as the eID and the counterValue (and other values in the future).

So let's create a master data entry the first time when a REST order
with a not yet known eID is made. For now we assume that the eUICC has
just been initialized with the eIM configuration and its counter is
at its minimum value. Whenever an eUICC package is generated, we
increment the counter by one and store the incremented value. This
will give us fresh counter values for each eUICC package.

Related: SYS#6720