esipa_asn1_handler: improve handling of provideEimPackageResult
This patch improves the handling of provideEimPackageResult. The
code that handles the euiccPackageResult is now in a separate file and
is improved so that it handles error responses better.

We are now also able to extract the eimPackageError code and communicate
it back via the REST API.

What still is missing though are signature checks and handling of
profileDownloadTriggerResult and ipaEuiccDataResponse

Related: SYS#6720