mnesia_db: add code to maintain a clean database
To maintain a clean database we should check if there any stuck
orders or orders that are not properly maintained anymore. This
is to prevent that the database fills up with orpaned items over
time.

The function that does the cleanup must be called in regular
intervals. This is not active yet and will be adressed in a follow
up patch.

Related: SYS#6720