crypto_utils: verification of signatures in eUICC package responses
This patch makes the code that verifies the signatures of the responses
to eUICC packages more complete. However, in practice the signature
verification still does not work. The reason is still unclear. This
patch adds some experimental code to prove that the cryptographic
aspects are correct. It is likely that the problem is the construction
of the message to be verified. (see also crypto_utils.erl).

This eventually means that this part of the code is still work in
progress.

Related: SYS#6720