esipa_asn1_handler: store eUICC public key during download
When a profile is downloaded the EUM and the eUICC certificate is
transferred to the eIM via authenticateClientRequestEsipa. In case
we do not know the eUICC public key yet (we need it to sign eUICC
packages), we can extract it from the eUICC certificate. However,
we also must make sure that the certificate chain is valid.

Related: SYS#6720